LPD-60239 move test to InvokerFilterTest
